Moving dialog data to separate file
